Removed inner module from generated code. Implemented send_component_update.
